requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
glcontext==2.5.0
idna==3.6
moderngl==5.10.0
numpy==1.26.4
pygame==2.5.2
PyGLM==2.7.1
PyWavefront==1.3.3
requests==2.31.0
urllib3==2.2.1
